requirements.txt
langchain>=0.2.13
langgraph>=0.0.45
langchain-openai>=0.1.22
requests>=2.31.0
python-dotenv>=1.0.1
jsonschema>=4.23.0
langchain-google-genai>=1.0.7
pydantic>=2.6.0